requirements.txt
amqp==2.1.1
billiard==3.5.0.2
celery==4.0.0
coreapi==1.32.0
coverage==4.2
defusedxml==0.4.1
Django==1.8.13
django-countries==3.4.1
django-filter==0.13.0
django-ipware==1.1.5
django-rest-swagger==0.3.10
djangorestframework==3.4.3
djangorestframework-xml==1.3.0
itypes==1.1.0
kombu==4.0.0
Markdown==2.6.6
mod-wsgi==4.5.5
mysqlclient==1.3.7
npTDMS==0.8.2
numpy==1.11.2
openapi-codec==1.0.0
pandas==0.19.1
pynmea2==1.5.3
python-dateutil==2.6.0
pytz==2016.7
PyYAML==3.11
requests==2.12.1
six==1.10.0
SQLAlchemy==1.1.4
uritemplate==3.0.0
vine==1.1.3
